code climate fix
